Create Dtos, mapper and update the requests methods

+++ src/main/java/com/example/BoatRegistry/services/BoatTypeService.java
@@ -1,6 +1,8 @@
 package com.example.BoatRegistry.services;
 
-import com.example.BoatRegistry.store.entities.BoatType;
+import com.example.BoatRegistry.store.dtos.BoatTypeRequestDto;
+import com.example.BoatRegistry.store.dtos.BoatTypeResponseDto;
+import com.example.BoatRegistry.store.mappers.BoatTypeMapper;
 import com.example.BoatRegistry.store.repositories.BoatTypeRepository;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.stereotype.Service;
@@ -11,36 +13,55 @@ import java.util.List;
 public class BoatTypeService {
 
     private final BoatTypeRepository boatTypeRepository;
+    private final BoatTypeMapper boatTypeMapper;
+    private final String NotFoundMessage = "Boat type with id %s not found";
 
-    public BoatTypeService(BoatTypeRepository boatTypeRepository) {
+    public BoatTypeService(BoatTypeRepository boatTypeRepository, BoatTypeMapper boatTypeMapper) {
         this.boatTypeRepository = boatTypeRepository;
+        this.boatTypeMapper = boatTypeMapper;
     }
 
-    public List<BoatType> getAll() {
-       return boatTypeRepository.findAll();
+    public List<BoatTypeResponseDto> getAll() {
+       var boatTypes = boatTypeRepository.findAll();
+       return boatTypeMapper.toResponseDtoList(boatTypes);
     }
 
-    public BoatType getById(Long id) {
-        return boatTypeRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("BoatType with id " + id + " not found"));
+    public BoatTypeResponseDto getById(Long id) {
+       var boatTypeOptional = boatTypeRepository.findById(id);
+
+       if(boatTypeOptional.isEmpty()) {
+           throw new EntityNotFoundException(String.format(NotFoundMessage, id));
+       }
+
+       var boatType = boatTypeOptional.get();
+       return boatTypeMapper.toResponseDto(boatType);
     }
 
-    public BoatType save(BoatType boatType) {
-        // validations here or in the controller ???
-        return boatTypeRepository.save(boatType);
+    public BoatTypeResponseDto save(BoatTypeRequestDto boatTypeRequestDto) {
+        var boatTypeToInsert = boatTypeMapper.toEntity(boatTypeRequestDto);
+        var insertedBoatType =  boatTypeRepository.save(boatTypeToInsert);
+        return boatTypeMapper.toResponseDto(insertedBoatType);
     }
 
-    public BoatType update(Long id, BoatType boatType) {
-        var boatTypeToUpdate = boatTypeRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("BoatType with id " + id + " not found"));
+    public BoatTypeResponseDto update(Long id, BoatTypeRequestDto boatTypeRequestDto) {
+        var boatTypeOptional = boatTypeRepository.findById(id);
 
-        boatTypeToUpdate.setName(boatType.getName());
-        return boatTypeRepository.save(boatTypeToUpdate);
+        if(boatTypeOptional.isEmpty()) {
+            throw new EntityNotFoundException(String.format(NotFoundMessage, id));
+        }
+        var boatType = boatTypeOptional.get();
+        boatType.setName(boatTypeRequestDto.getName());
+        var updatedBoatType = boatTypeRepository.save(boatType);
+        return boatTypeMapper.toResponseDto(updatedBoatType);
     }
 
     public void delete(Long id) {
-        var boatType = boatTypeRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("BoatType with id " + id + " not found"));
+        var boatTypeOptional = boatTypeRepository.findById(id);
+
+        if(boatTypeOptional.isEmpty()) {
+            throw new EntityNotFoundException(String.format(NotFoundMessage, id));
+        }
+        var boatType = boatTypeOptional.get();
         boatTypeRepository.delete(boatType);
     }
 }
